+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Fixed
+- `Supermetrics` source: `to_df()` now correctly handles `if_empty` in case of empty results
 
 ## [0.2.12]
 ### Added